// src/types.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  USER = 'USER',
}

export type User = {
  id: string
  username: string
  name: string
  email: string
  image: string
  token: string
  salt: string
  role: UserRole
  chipsAmount: number
  createdAt: Date
  updatedAt: Date
}

export type Card = {
  id: string
  rank: string
  suit: string
}

export type History = {
  id: string

  content: string
  deleted: boolean
  tableId: string
  table?: Table

  playerId: string
  player?: Player

  createdAt: Date
}

export type Table = {
  id: string
  name: string
  eventId: string
  event?: Event

  userId: string
  user?: User

  players: Player[]
  messages: Message[]
  matches?: Match[]
  handOver: boolean
  chatBanned: boolean

  minBuyIn: number
  maxBuyIn: number
  ante: number
  maxPlayers: number

  histories: History[]

  createdAt: Date
  updatedAt: Date
}

export type Player = {
  id: string

  userId: string
  user?: User

  tableId: string
  table?: Table
  stack: number
  buyIn: number
  socketId: string

  isTurn: boolean
  leaveNextMatch: boolean

  createdAt: Date
}

export type Message = {
  id: string
  content: string
  stickerImageSrc: string

  playerId: string
  player?: Player

  tableId: string
  table?: Table

  deleted: boolean
  createdAt: Date
}

export type Deck = {
  id: string
  tableId: string
  table: Table
  cards?: Card[]
  matches?: Match[]
}

export type WinMessages = {
  id: string
  amount: string
  matchId: string
  match?: Match
  handName: string
  bestHand?: string
  winnerHand?: string
  content: string
  createdAt: Date
}

export type Match = {
  id: string
  tableId: string
  table: Table
  numberPlayers: number
  deckId: string
  deck: Deck
  participants: Participant[]
  board: Card[]
  isPreFlop: boolean
  isFlop: boolean
  isTurn: boolean
  isRiver: boolean
  isShowdown: boolean
  isAllAllIn: boolean
  winners?: Player[]
  buttonId?: string
  button?: Player
  bigBlindId?: string
  bigBlind?: Player
  smallBlindId?: string
  smallBlind?: Player
  callAmount: number
  minRaise: number
  minBet: number
  pot: number
  winMessages?: WinMessages[]
}